requirements.txt
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.9.1
binaryornot==0.4.4
cachetools==5.5.2
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
dj-database-url==3.0.1
Django==5.2.4
django-tailwind==4.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
